services/ai_processor/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-genai==1.7.0
pydantic==2.6.4
